Improved cooldown data
Added a cooldown id enum for skill and cooldown/overheat data to use and renamed a few fields to make their purpose more clear.

+++ src/ZoneServer/Network/Send.cs
@@ -464,7 +464,7 @@ namespace Melia.Zone.Network
 			var packet = new Packet(Op.ZC_OVERHEAT_CHANGED);
 
 			packet.PutLong(character.Id);
-			packet.PutInt(skill.OverheatData.Id);
+			packet.PutInt((int)skill.OverheatData.Id);
 			packet.PutInt(skill.OverheatCounter * skill.OverheatData.OverheatResetTime);
 			packet.PutInt(0);
 			packet.PutInt(skill.OverheatData.OverheatResetTime);
